refactor: change flow graph theme on runner page
