+++ apps/web-app/src/components/RunnerFlowSelector.tsx
@@ -1,18 +1,26 @@
 import { useAppSelector } from "../redux/typed-hooks";
-import { useGetFlowDefQuery, useGetFlowsQuery } from "../redux/api/flows-api";
+import { useGetFlowsQuery } from "../redux/api/flows-api";
 import { useRequestRunMutation } from "../redux/api/runs-api";
 import { useDispatch } from "react-redux";
 import { addEventPanel, setFlowSelectedId } from "../redux/slices/runner-slice";
-import { skipToken } from "@reduxjs/toolkit/query";
-import { RunnerFlowView } from "./RunnerFlowView";
+import { Button } from "./ui/button";
+
+import {
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectTrigger,
+  SelectValue,
+} from "./ui/select";
 
 export function RunnerFlowSelector() {
   const { data } = useGetFlowsQuery();
   const dispatch = useDispatch();
   // const flowHash = useAppSelector((state) => state.runner.flowHash);
-  const [requestRun, requestRunState] = useRequestRunMutation();
+  const [requestRun] = useRequestRunMutation();
   const flowSelectedId = useAppSelector((state) => state.runner.flowSelectedId);
-  const flowDefQuery = useGetFlowDefQuery(flowSelectedId ?? skipToken);
 
   const handleRun = async () => {
     if (!flowSelectedId || !data?.ok) return;
@@ -23,37 +31,38 @@ export function RunnerFlowSelector() {
     }
   };
   return (
-    <div>
-      <label className="block">Select A Flow Definition:</label>
-      <select
-        className="bg-slate-700 p-2 min-w-50 cursor-pointer"
-        onChange={(e) => {
-          dispatch(setFlowSelectedId(e.target.value));
+    <div className="flex cursor-pointer">
+      <Select
+        onValueChange={(value) => {
+          console.log(value);
+          dispatch(setFlowSelectedId(value));
         }}
-        value={flowSelectedId ?? ""}
       >
-        <option value="">-- Select a Flow --</option>
-        {data?.ok === true
-          ? data.indexes.map((index) => (
-              <option value={index.hash} key={index.hash}>
-                {index.name} - {index.version}
-              </option>
-            ))
-          : ""}
-      </select>
-      <button
-        className="ml-3"
+        <SelectTrigger className="w-full max-w-100">
+          <SelectValue placeholder="Select A Flow" />
+        </SelectTrigger>
+        <SelectContent>
+          <SelectGroup>
+            <SelectLabel>Select A Flow</SelectLabel>
+            {data?.ok === true
+              ? data.indexes.map((index) => (
+                  <SelectItem value={index.hash} key={index.hash}>
+                    {index.name} - {index.version}
+                  </SelectItem>
+                ))
+              : ""}
+          </SelectGroup>
+        </SelectContent>
+      </Select>
+      <Button
+        className="ml-3 cursor-pointer"
+        variant="outline"
         onClick={handleRun}
         disabled={flowSelectedId === null || flowSelectedId === ""}
+        style={{}}
       >
         Run
-      </button>
-      <span>{requestRunState.isLoading ? "loading" : ""}</span>
-      {flowDefQuery.data?.ok ? (
-        <RunnerFlowView flowDef={flowDefQuery.data.value} />
-      ) : (
-        ""
-      )}
+      </Button>
     </div>
   );
 }
