+++ apps/web-app/src/index.css
@@ -8,9 +8,9 @@
 
 :root {
   --radius: 0.625rem;
-  --background: oklch(0.94 0 0);
+  --background: oklch(0.91 0 0);
   --foreground: oklch(0.3 0 0);
-  --card: oklch(1 0 0);
+  --card: oklch(0.8 0 0);
   --card-foreground: oklch(0.145 0 0);
   --popover: oklch(1 0 0);
   --primary: oklch(0.205 0 0);
@@ -24,8 +24,8 @@
   --accent-foreground: oklch(0.205 0 0);
   --destructive: oklch(0.577 0.245 27.325);
   --destructive-foreground: oklch(0.985 0 0);
-  --border: oklch(0.922 0 0);
-  --input: oklch(0.922 0 0);
+  --border: oklch(0.2 0 0);
+  --input: oklch(0.3 0 0);
   --ring: oklch(0.708 0 0);
   --chart-1: oklch(0.646 0.222 41.116);
   --chart-2: oklch(0.6 0.118 184.704);
@@ -43,7 +43,7 @@
 }
 
 .dark {
-  --background: oklch(0.25 0 0);
+  --background: oklch(19% 0 0);
   --foreground: oklch(0.94 0 0);
   --card: oklch(0.205 0 0);
   --card-foreground: oklch(0.985 0 0);
