+++ apps/web-app/src/layout/Header.tsx
@@ -1,11 +1,9 @@
-import { ModeToggle } from "@/components/mode-toggle";
 import { Nav } from "./Nav";
 
 export function Header() {
   return (
-    <header className="flex items-center gap-4 border-b border-border bg-background text-foreground px-3 py-2">
-      <h2 className="font-bold text-md">lowercase</h2>
-      <ModeToggle />
+    <header className="flex justify-between gap-4 border-b border-border bg-background text-foreground px-3 py-2">
+      <h2 className="font-bold text-md ml-0 pl-0">lowercase</h2>
       <Nav />
     </header>
   );
